feat&refactor: add empty files for settings page, rename settingMenu to settingsPopap
