+++ src/components/ForSettingsPage/SettingsList/SettingsList.tsx
@@ -1,0 +1,9 @@
+import './SettingsList.scss';
+
+function SettingsList() {
+  return (
+    <div className="user-info">This is SettingsList</div>
+  );
+}
+
+export default SettingsList;
