+++ src/components/ForSettingsPage/SettingsTopPanel/SettingsTopPanel.tsx
@@ -1,0 +1,9 @@
+import './SettingsTopPanel.scss';
+
+function SettingsTopPanel() {
+  return (
+    <div className="user-info">This is SettingsTopPanel</div>
+  );
+}
+
+export default SettingsTopPanel;
