+++ src/components/SettingsPopap/SettingsPopap.scss
